refactor: UI 개선 및 폴더 구조 정리
