+++ src/components/property/PropertyCard.tsx
